Rework Promise type
